--- a/weasyprint/formatting_structure/boxes.py
+++ b/weasyprint/formatting_structure/boxes.py
@@ -312,10 +312,9 @@ class ParentBox(Box):
         if end:
             self._reset_spacing('bottom')
 
-    def copy_with_children(self, new_children, is_start=True, is_end=True,
-                           copy_style=False):
+    def copy_with_children(self, new_children, is_start=True, is_end=True):
         """Create a new equivalent box with given ``new_children``."""
-        new_box = self.copy(copy_style=copy_style)
+        new_box = self.copy()
         new_box.children = tuple(new_children)
         if not is_start:
             new_box.outside_list_marker = None