--- a/weasyprint/formatting_structure/build.py
+++ b/weasyprint/formatting_structure/build.py
@@ -173,7 +173,8 @@ def element_to_box(element, style_for, get_image_from_uri, state=None):
         if not counter_values[name]:
             counter_values.pop(name)
 
-    box = box.copy_with_children(children)
+    if children:
+        box = box.copy_with_children(children)
     replace_content_lists(element, box, style, counter_values)
 
     # Specific handling for the element. (eg. replaced element)