--- a/weasyprint/layout/__init__.py
+++ b/weasyprint/layout/__init__.py
@@ -57,7 +57,7 @@ def layout_document(enable_hinting, style_for, get_image_from_uri, root_box):
         root_children.extend(layout_fixed_boxes(context, pages[:i]))
         root_children.extend(root.children)
         root_children.extend(layout_fixed_boxes(context, pages[i+1:]))
-        root = root.copy_with_children(root_children)
+        root.children = root_children
         context.current_page = page_counter[0]
         page.children = (root,) + tuple(
             make_margin_boxes(context, page, counter_values))