--- a/weasyprint/layout/blocks.py
+++ b/weasyprint/layout/blocks.py
@@ -389,7 +389,7 @@ def block_container_layout(context, box, max_position_y, skip_stack,
     is_start = skip_stack is None
     if not is_start:
         # Remove top margin, border and padding:
-        box = box.copy_with_children(box.children, is_start=False)
+        box._remove_decoration(start=True, end=False)
 
     if adjoining_margins is None:
         adjoining_margins = []
@@ -719,8 +719,7 @@ def block_container_layout(context, box, max_position_y, skip_stack,
         adjoining_margins = []
 
     new_box = box.copy_with_children(
-        new_children, is_start=is_start, is_end=resume_at is None,
-        copy_style=True)
+        new_children, is_start=is_start, is_end=resume_at is None)
 
     # TODO: See corner cases in
     # http://www.w3.org/TR/CSS21/visudet.html#normal-block
@@ -847,8 +846,7 @@ def find_earlier_page_break(children, absolute_boxes, fixed_boxes):
                     child.children, absolute_boxes, fixed_boxes)
                 if result:
                     new_grand_children, resume_at = result
-                    new_child = child.copy_with_children(
-                        new_grand_children, copy_style=True)
+                    new_child = child.copy_with_children(new_grand_children)
                     new_children = list(children[:index]) + [new_child]
                     # Index in the original parent
                     resume_at = (new_child.index, resume_at)