--- a/weasyprint/layout/inlines.py
+++ b/weasyprint/layout/inlines.py
@@ -155,8 +155,7 @@ def get_next_linebox(context, linebox, position_y, skip_stack,
             absolute_boxes, fixed_boxes)
         float_children.append(waiting_float)
     if float_children:
-        line = line.copy_with_children(
-            line.children + tuple(float_children))
+        line.children += tuple(float_children)
 
     return line, resume_at
 
@@ -627,7 +626,7 @@ def split_inline_box(context, box, position_x, max_x, skip_stack,
         resume_at = None
 
     new_box = box.copy_with_children(
-        children, is_start=is_start, is_end=resume_at is None, copy_style=True)
+        children, is_start=is_start, is_end=resume_at is None)
     if isinstance(box, boxes.LineBox):
         # Line boxes already have a position_x which may not be the same
         # as content_box_left when text-indent is non-zero.