--- a/weasyprint/layout/pages.py
+++ b/weasyprint/layout/pages.py
@@ -479,7 +479,7 @@ def make_page(context, root_box, page_type, resume_at, content_empty,
     """
     style = context.style_for(page_type)
     # Propagated from the root or <body>.
-    style.overflow = root_box.viewport_overflow
+    style['overflow'] = root_box.viewport_overflow
     page = boxes.PageBox(page_type, style)
 
     device_size = page.style.size