--- a/weasyprint/stacking.py
+++ b/weasyprint/stacking.py
@@ -138,7 +138,8 @@ class StackingContext(object):
                 result = dispatch(child)
                 if result is not None:
                     new_children.append(result)
-            return box.copy_with_children(new_children)
+            box.children = new_children
+            return box
 
         box = dispatch_children(box)