… weasyprint/tests/test_boxes.py
… weasyprint/tests/test_presentational_hints.py
… weasyprint/tests/test_text.py