The main point of this pull request is to make `StyleDict` inherit from `dict` instead of the copy-on-write-like strategy used before.

This change really improves the speed (yeah, profiling helps), but unfortunately has a high memory footprint. That's why memory usage has been also improved, including an impressive don't-store-the-dummy-cairo-context-in-the-layout-forever in 152193c (already incluled in the `font-face` branch).

The commits are small and hopefully easy to understand and to revert if needed. I've got exactly the same PDFs before and after these commits with many examples, but I'm sure that bugs have been introduced.

Here are the results (with reliable science inside):

## Offline HTML5 Specification

https://www.w3.org/TR/html5/

Long document with a lot of lists and underlined links.

0.31:
- 6.0s
- 146MB memory peak

font-face:
- 6.3s (+5%)
- 130MB memory peak (-11%)

font-face + perf:
- 4.5s (-25%)
- 159MB memory peak (+9%)

## Offline Alice Adventures in Wonderland

https://www.gutenberg.org/files/11/11-h/11-h.htm

Long document with repetitive justified text.

0.31:
- 12.5s
- 162MB memory peak

font-face:
- 9.8s (-21%)
- 134MB memory peak (-17%)

font-face + perf:
- 7.1s (-43%, **too good to be true**)
- 161MB memory peak (=)

## Online WeasyPrint

http://weasyprint.org/

Simple website, uses online fonts (that's why it's slower now).

0.31:
- 1.6s
- 66MB memory peak

font-face:
- 2.2s (+38%)
- 68MB memory peak (+3%)

font-face + perf:
- 2.0s (+25%)
- 70MB memory peak (+6%)

## Online Wikipedia

https://en.wikipedia.org/w/index.php?title=HTML5&printable=yes

Printable version of a Wikipedia page, long left-aligned paragraphs with floats.

0.31:
- 7.9s
- 134MB memory peak

font-face:
- 8.2s (+4%)
- 124MB memory peak (-7%)

font-face + perf:
- 6.6s (-16%)
- 141MB memory peak (+5%)