@liZe 2016-11-10
I've fixed most of the memory leaks, there's no crash with the W3C test suites anymore, let's merge.